A steady state commit.The language is usable.

+++ build_and_run.sh
@@ -6,8 +6,19 @@ JAR_NAME="hecate.jar"
 MAIN_CLASS="com.babel.hecate.Hecate"
 PARSE_FILE_NAME="./tests/parsertest-class.txt"
 
+
+
+if [[ $# -ne 1 ]] ; then
+    echo 'Pass a text file with the source code as an argument. Only one argument'
+    exit 0
+fi
+
+
 echo "Removing old binary files in ${OUT}/*"
 rm -rf "${OUT}"/*
+
+
+
 
 echo "Compiling"
 
@@ -20,6 +31,8 @@ fi
 
 jar cfm "${OUT}/${JAR_NAME}" MANIFEST.MF -C bin .
 
+
+
 if [[ $? -ne 0 ]]; then
     echo "Failed to create jar file!"
     exit 1
@@ -30,7 +43,7 @@ echo "Generated artifact ${JAR_NAME} in ${OUT}"
 
 echo "Running artifact"
 
-java -jar "${OUT}/${JAR_NAME}" "${PARSE_FILE_NAME}"
+java -jar "${OUT}/${JAR_NAME}" "$1"
 
 if [[ $? -ne 0 ]]; then
     echo "Failed to run the jar file!"
